refactor: logout 로직 위치 수정
